LJ-4 added sql functions for adding entities into database and getting information from database

+++ src/main/resources/database/postgres/function/filter_trips_by_cost.sql
@@ -1,4 +1,4 @@
-CREATE OR REPLACE FUNCTION filter_by_cost(minCost numeric, maxCost numeric) RETURNS SETOF trip
+CREATE OR REPLACE FUNCTION filter_by_cost(minCost double precision, maxCost double precision) RETURNS SETOF trip
     LANGUAGE SQL
 AS
 $$
